+++ engines/gob/save/saveload_v7.cpp
@@ -101,7 +101,7 @@ SaveLoad_v7::~SaveLoad_v7() {
 const SaveLoad_v7::SaveFile *SaveLoad_v7::getSaveFile(const char *fileName) const {
 	fileName = stripPath(fileName);
 
-	for (int i = 0; i < ARRAYSIZE(_saveFiles); i++)
+	for (size_t i = 0; i < ARRAYSIZE(_saveFiles); i++)
 		if (!scumm_stricmp(fileName, _saveFiles[i].sourceName))
 			return &_saveFiles[i];
 
@@ -111,7 +111,7 @@ const SaveLoad_v7::SaveFile *SaveLoad_v7::getSaveFile(const char *fileName) cons
 SaveLoad_v7::SaveFile *SaveLoad_v7::getSaveFile(const char *fileName) {
 	fileName = stripPath(fileName);
 
-	for (int i = 0; i < ARRAYSIZE(_saveFiles); i++)
+	for (size_t i = 0; i < ARRAYSIZE(_saveFiles); i++)
 		if (!scumm_stricmp(fileName, _saveFiles[i].sourceName))
 			return &_saveFiles[i];
