+++ engines/kyra/kyra_v2.cpp
@@ -134,7 +134,7 @@ void KyraEngine_v2::pauseEngineIntern(bool pause) {
 	if (!pause) {
 		uint32 pausedTime = _system->getMillis() - _pauseStart;
 
-		for (int i = 0; i < ARRAYSIZE(_sceneSpecialScriptsTimer); ++i) {
+		for (size_t i = 0; i < ARRAYSIZE(_sceneSpecialScriptsTimer); ++i) {
 			if (_sceneSpecialScriptsTimer[i])
 				_sceneSpecialScriptsTimer[i] += pausedTime;
 		}
