+++ engines/kyra/scene_hof.cpp
@@ -681,7 +681,7 @@ void KyraEngine_HoF::initSceneScreen(int unk1) {
 }
 
 void KyraEngine_HoF::freeSceneShapePtrs() {
-	for (int i = 0; i < ARRAYSIZE(_sceneShapeTable); ++i)
+	for (size_t i = 0; i < ARRAYSIZE(_sceneShapeTable); ++i)
 		delete[] _sceneShapeTable[i];
 	memset(_sceneShapeTable, 0, sizeof(_sceneShapeTable));
 }
