+++ engines/scumm/usage_bits.cpp
@@ -42,21 +42,21 @@ void ScummEngine::upgradeGfxUsageBits() {
 }
 
 void ScummEngine::setGfxUsageBit(int strip, int bit) {
-	assert(strip >= 0 && strip < ARRAYSIZE(gfxUsageBits) / 3);
+	assert(strip >= 0 && (size_t)strip < ARRAYSIZE(gfxUsageBits) / 3);
 	assert(1 <= bit && bit <= 96);
 	bit--;
 	gfxUsageBits[3 * strip + bit / 32] |= (1 << (bit % 32));
 }
 
 void ScummEngine::clearGfxUsageBit(int strip, int bit) {
-	assert(strip >= 0 && strip < ARRAYSIZE(gfxUsageBits) / 3);
+	assert(strip >= 0 && (size_t)strip < ARRAYSIZE(gfxUsageBits) / 3);
 	assert(1 <= bit && bit <= 96);
 	bit--;
 	gfxUsageBits[3 * strip + bit / 32] &= ~(1 << (bit % 32));
 }
 
 bool ScummEngine::testGfxUsageBit(int strip, int bit) {
-	assert(strip >= 0 && strip < ARRAYSIZE(gfxUsageBits) / 3);
+	assert(strip >= 0 && (size_t)strip < ARRAYSIZE(gfxUsageBits) / 3);
 	assert(1 <= bit && bit <= 96);
 	bit--;
 	return (gfxUsageBits[3 * strip + bit / 32] & (1 << (bit % 32))) != 0;
@@ -67,7 +67,7 @@ bool ScummEngine::testGfxAnyUsageBits(int strip) {
 	uint32 bitmask[3] = { 0xFFFFFFFF, 0xFFFFFFFF, 0x3FFFFFFF };
 	int i;
 
-	assert(strip >= 0 && strip < ARRAYSIZE(gfxUsageBits) / 3);
+	assert(strip >= 0 && (size_t)strip < ARRAYSIZE(gfxUsageBits) / 3);
 	for (i = 0; i < 3; i++)
 		if (gfxUsageBits[3 * strip + i] & bitmask[i])
 			return true;
@@ -80,7 +80,7 @@ bool ScummEngine::testGfxOtherUsageBits(int strip, int bit) {
 	uint32 bitmask[3] = { 0xFFFFFFFF, 0xFFFFFFFF, 0xFFFFFFFF };
 	int i;
 
-	assert(strip >= 0 && strip < ARRAYSIZE(gfxUsageBits) / 3);
+	assert(strip >= 0 && (size_t)strip < ARRAYSIZE(gfxUsageBits) / 3);
 	assert(1 <= bit && bit <= 96);
 	bit--;
 	bitmask[bit / 32] &= ~(1 << (bit % 32));
